+++ quest_test/test_manager.py
@@ -36,7 +36,7 @@ async def test_manager():
 
     async with WorkflowManager('test-manager', storage, create_history, lambda w_type: workflow,
                                serializer=NoopSerializer()) as manager:
-        manager.start_workflow('workflow', 'wid1', 4)
+        manager.start_workflow('workflow', 'wid1', 4, delete_on_finish=False)
         await asyncio.sleep(0.1)
         # Now pause the manager and all workflows
 
@@ -49,7 +49,7 @@ async def test_manager():
         # At this point, all workflows should be resumed
         pause.set()
         await asyncio.sleep(0.1)
-        result = await manager.get_workflow('wid1')
+        result = await manager.get_workflow_result('wid1')
         assert result == 11
 
     assert counter_a == 2
@@ -89,7 +89,7 @@ async def test_manager_events():
 
     async with WorkflowManager('test-manager', storage, create_history, lambda w_type: workflow,
                                serializer=NoopSerializer()) as manager:
-        manager.start_workflow('workflow', 'wid1', 1)
+        manager.start_workflow('workflow', 'wid1', 1, delete_on_finish=False)
         await asyncio.sleep(0.1)
         await manager.send_event('wid1', 'messages', None, 'put', 2)
         await asyncio.sleep(0.1)
@@ -105,7 +105,7 @@ async def test_manager_events():
         await asyncio.sleep(0.1)
         await manager.send_event('wid1', 'messages', None, 'put', 3)
         await manager.send_event('wid1', 'messages', None, 'put', 0)  # i.e. end the workflow
-        result = await manager.get_workflow('wid1')
+        result = await manager.get_workflow_result('wid1')
         assert result == 6
 
     assert counter_a == 2
@@ -146,7 +146,7 @@ async def test_manager_background():
 
     async with WorkflowManager('test-manager', storage, create_history, lambda w_type: workflow,
                                serializer=NoopSerializer()) as manager:
-        manager.start_workflow_background('workflow', 'wid1', 1)
+        manager.start_workflow('workflow', 'wid1', 1)
         await asyncio.sleep(0.1)
         await manager.send_event('wid1', 'messages', None, 'put', 2)
         await asyncio.sleep(0.1)
@@ -188,7 +188,7 @@ async def test_get_queue():
 
     async with WorkflowManager('test', storage, create_history, lambda wid: workflow,
                                serializer=NoopSerializer()) as wm:
-        wm.start_workflow('workflow', 'wid')
+        wm.start_workflow('workflow', 'wid', delete_on_finish=False)
         await asyncio.sleep(0.1)
         q = await wm.get_queue('wid', 'messages', None)
         result = await wm.get_state('wid', 'result', None)